Sync of some (yet) untested work (Host Cache) * Refactored hosts storage and eliminated usage of the vector. * Renamed CHostCache to CG2HostCache (including related variables and typedefs). * Refactored G2Packet::WriteHostAddress to use a const reference instead of a pointer. (fixes some gcc warning) * Constified CNeighboursBase::Find (another gcc warning) * Changed Connection.FailurePenalty and Gnutella.ConnectThrottle to quint16

+++ Quazaa/quazaasettings.h
@@ -163,7 +163,7 @@ namespace Settings
 		bool		DetectConnectionLoss;					// Detect loss of internet
 		bool		DetectConnectionReset;					// Detect regaining of internet connection
 		quint8		FailureLimit;							// Max allowed connection failures (default = 3) (Neighbour connections)
-		int			FailurePenalty;							// Delay after connection failure (seconds, default = 300) (Neighbour connections)
+		quint16		FailurePenalty;							// Delay after connection failure (seconds, default = 300) (Neighbour connections)
 		quint64		InSpeed;								// Inbound internet connection speed in B/s
 		quint64		OutSpeed;								// Outbound internet connection speed in B/s
 		quint16		Port;									// Incoming port
@@ -279,7 +279,7 @@ namespace Settings
 	struct sGnutella
 	{
 		int			ConnectFactor;
-		quint32		ConnectThrottle;						// Delay between connection attempts (seconds)
+		quint16		ConnectThrottle;						// Delay between connection attempts (seconds)
 		int			HitsPerPacket;							// Max hits allowed for a single packet
 		quint32		HostCacheSize;							// Max size of host cache; 0 to disable
 		int			HostCacheView;
